requirements.txt
streamlit
pillow
numpy
python-dotenv
landingai
